# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "diagnostic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostic Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "retryable",
          "diagnostic_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "The one error shape, for the generated contract (the handlers below produce it).",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ForgetPending": {
        "description": "Forgetting happened; part of the physical cleanup could not finish and is retried.",
        "properties": {
          "outstanding": {
            "items": {
              "type": "string"
            },
            "title": "Outstanding",
            "type": "array"
          }
        },
        "required": [
          "outstanding"
        ],
        "title": "ForgetPending",
        "type": "object"
      },
      "IdentityLink": {
        "properties": {
          "identity_id": {
            "format": "uuid",
            "title": "Identity Id",
            "type": "string"
          }
        },
        "required": [
          "identity_id"
        ],
        "title": "IdentityLink",
        "type": "object"
      },
...
      "NamePerson": {
        "properties": {
          "display_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "identity_id": {
            "format": "uuid",
            "title": "Identity Id",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "identity_id"
        ],
        "title": "NamePerson",
        "type": "object"
      },
...
      "PageInfo": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "next_cursor",
          "has_more"
        ],
        "title": "PageInfo",
        "type": "object"
      },
...
      "Page_PersonSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PersonSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "Page[PersonSummary]",
        "type": "object"
      },
...
      "PersonSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "identity_count": {
            "title": "Identity Count",
            "type": "integer"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "revision",
          "identity_count",
          "created_at",
          "updated_at"
        ],
        "title": "PersonSummary",
        "type": "object"
      },
...
      "RenamePerson": {
        "properties": {
          "display_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "display_name",
          "expected_revision"
        ],
        "title": "RenamePerson",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FaceIdentify local API",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/people": {
      "get": {
        "operationId": "list_people",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PersonSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The one error shape."
          }
        },
        "summary": "List People",
        "tags": [
          "people"
        ]
      },
      "post": {
        "operationId": "name_an_identity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamePerson"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The one error shape."
          }
        },
        "summary": "Name An Identity",
        "tags": [
          "people"
        ]
      }
    },
    "/api/v1/people/{person_id}": {
      "get": {
        "operationId": "get_person",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The one error shape."
          }
        },
        "summary": "Get Person",
        "tags": [
          "people"
        ]
      },
      "patch": {
        "operationId": "rename",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenamePerson"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The one error shape."
          }
        },
        "summary": "Rename",
        "tags": [
          "people"
        ]
      }
    },
    "/api/v1/people/{person_id}/assign-identity": {
      "post": {
        "operationId": "assign_identity",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityLink"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The one error shape."
          }
        },
        "summary": "Assign Identity",
        "tags": [
          "people"
        ]
      }
    },
    "/api/v1/people/{person_id}/forget": {
      "post": {
        "description": "Forget a person (identity model section 46): run `ForgetIdentity` on each identity of the\nPerson. The Person, their name and notes stay, without visual support. A repeat is harmless.",
        "operationId": "forget_person",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgetPending"
                }
              }
            },
            "description": "Accepted"
          },
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The one error shape."
          }
        },
        "summary": "Forget Person",
        "tags": [
          "people"
        ]
      }
    },
    "/api/v1/people/{person_id}/remove-identity": {
      "post": {
        "operationId": "remove_identity",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityLink"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The one error shape."
          }
        },
        "summary": "Remove Identity",
        "tags": [
          "people"
        ]
      }
    },
...
  }
}